+++ polars/polars-lazy/src/physical_plan/expressions/default.rs
@@ -9,33 +9,6 @@ use rayon::prelude::*;
 use std::borrow::Cow;
 use std::sync::Arc;
 
-pub struct NotExpr(Arc<dyn PhysicalExpr>, Expr);
-
-impl NotExpr {
-    pub fn new(physical_expr: Arc<dyn PhysicalExpr>, expr: Expr) -> Self {
-        Self(physical_expr, expr)
-    }
-}
-impl PhysicalExpr for NotExpr {
-    fn as_expression(&self) -> &Expr {
-        &self.1
-    }
-
-    fn evaluate(&self, df: &DataFrame, state: &ExecutionState) -> Result<Series> {
-        let series = self.0.evaluate(df, state)?;
-        if let Ok(ca) = series.bool() {
-            Ok((!ca).into_series())
-        } else {
-            Err(PolarsError::InvalidOperation(
-                format!("NotExpr expected a boolean type, got: {:?}", series).into(),
-            ))
-        }
-    }
-    fn to_field(&self, _input_schema: &Schema) -> Result<Field> {
-        Ok(Field::new("not", DataType::Boolean))
-    }
-}
-
 pub struct AliasExpr {
     pub(crate) physical_expr: Arc<dyn PhysicalExpr>,
     pub(crate) name: Arc<String>,
